+++ efb/src/measurements/constants.rs
@@ -18,6 +18,7 @@ pub const INCH_IN_METER: f32 = 0.0254;
 pub const KELVIN_IN_CELSIUS: f32 = 273.15;
 pub const METER_PER_SECONDS_IN_KNOTS: f32 = 1.943844;
 pub const NAUTICAL_MILE_IN_METER: f32 = 1852.0;
+pub const STATUTE_MILE_IN_METER: f32 = 1609.344;
 pub const PI2: f32 = std::f32::consts::PI * 2.0;
 pub const POUNDS_IN_KILOGRAMS: f32 = 0.4535924;
 pub const US_GALLON_IN_QUBIC_METER: f32 = 0.003785412;