+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Length in cm, km and statue miles
 - Volume in US gallon
 - Physical quantity an unit measures